+++ crates/ruma-common/src/events/room/message.rs
@@ -9,6 +9,7 @@ use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use serde_json::Value as JsonValue;
 
 use crate::{
+    events::relation::{InReplyTo, Replacement, Thread},
     serde::{JsonObject, StringEnum},
     OwnedEventId, PrivOwnedStr,
 };
@@ -589,77 +590,6 @@ pub enum Relation<C> {
 
     #[doc(hidden)]
     _Custom,
-}
-
-/// Information about the event a "rich reply" is replying to.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-pub struct InReplyTo {
-    /// The event being replied to.
-    pub event_id: OwnedEventId,
-}
-
-impl InReplyTo {
-    /// Creates a new `InReplyTo` with the given event ID.
-    pub fn new(event_id: OwnedEventId) -> Self {
-        Self { event_id }
-    }
-}
-
-/// The event this relation belongs to [replaces another event].
-///
-/// [replaces another event]: https://spec.matrix.org/v1.4/client-server-api/#event-replacements
-#[derive(Clone, Debug)]
-#[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-pub struct Replacement<C> {
-    /// The ID of the event being replaced.
-    pub event_id: OwnedEventId,
-
-    /// New content.
-    pub new_content: C,
-}
-
-impl<C> Replacement<C> {
-    /// Creates a new `Replacement` with the given event ID and new content.
-    pub fn new(event_id: OwnedEventId, new_content: C) -> Self {
-        Self { event_id, new_content }
-    }
-}
-
-/// The content of a thread relation.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-pub struct Thread {
-    /// The ID of the root message in the thread.
-    pub event_id: OwnedEventId,
-
-    /// A reply relation.
-    ///
-    /// If this event is a reply and belongs to a thread, this points to the message that is being
-    /// replied to, and `is_falling_back` must be set to `false`.
-    ///
-    /// If this event is not a reply, this is used as a fallback mechanism for clients that do not
-    /// support threads. This should point to the latest message-like event in the thread and
-    /// `is_falling_back` must be set to `true`.
-    pub in_reply_to: InReplyTo,
-
-    /// Whether the `m.in_reply_to` field is a fallback for older clients or a genuine reply in a
-    /// thread.
-    pub is_falling_back: bool,
-}
-
-impl Thread {
-    /// Convenience method to create a regular `Thread` with the given event ID and latest
-    /// message-like event ID.
-    pub fn plain(event_id: OwnedEventId, latest_event_id: OwnedEventId) -> Self {
-        Self { event_id, in_reply_to: InReplyTo::new(latest_event_id), is_falling_back: true }
-    }
-
-    /// Convenience method to create a reply `Thread` with the given event ID and replied-to event
-    /// ID.
-    pub fn reply(event_id: OwnedEventId, reply_to_event_id: OwnedEventId) -> Self {
-        Self { event_id, in_reply_to: InReplyTo::new(reply_to_event_id), is_falling_back: false }
-    }
 }
 
 /// The format for the formatted representation of a message body.
